+++ modules/features/account/src/main/java/au/com/shiftyjelly/pocketcasts/account/deviceapprove/DeviceApproveViewModel.kt
@@ -3,7 +3,14 @@ package au.com.shiftyjelly.pocketcasts.account.deviceapprove
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import au.com.shiftyjelly.pocketcasts.preferences.Settings
+import au.com.shiftyjelly.pocketcasts.repositories.playback.PlaybackManager
+import au.com.shiftyjelly.pocketcasts.repositories.playback.UpNextQueue
+import au.com.shiftyjelly.pocketcasts.repositories.podcast.EpisodeManager
+import au.com.shiftyjelly.pocketcasts.repositories.podcast.FolderManager
+import au.com.shiftyjelly.pocketcasts.repositories.searchhistory.SearchHistoryManager
 import au.com.shiftyjelly.pocketcasts.repositories.sync.SyncManager
+import au.com.shiftyjelly.pocketcasts.repositories.user.UserManager
+import au.com.shiftyjelly.pocketcasts.utils.log.LogBuffer
 import com.automattic.eventhorizon.DeviceApproveConnectTappedEvent
 import com.automattic.eventhorizon.DeviceApproveDismissedEvent
 import com.automattic.eventhorizon.DeviceApproveFailedEvent
@@ -25,6 +32,12 @@ import timber.log.Timber
 @HiltViewModel
 class DeviceApproveViewModel @Inject constructor(
     private val syncManager: SyncManager,
+    private val userManager: UserManager,
+    private val playbackManager: PlaybackManager,
+    private val upNextQueue: UpNextQueue,
+    private val folderManager: FolderManager,
+    private val searchHistoryManager: SearchHistoryManager,
+    private val episodeManager: EpisodeManager,
     private val settings: Settings,
     private val eventHorizon: EventHorizon,
 ) : ViewModel() {
@@ -32,7 +45,7 @@ class DeviceApproveViewModel @Inject constructor(
     private val _uiState = MutableStateFlow(DeviceApproveUiState())
     val uiState: StateFlow<DeviceApproveUiState> = _uiState.asStateFlow()
 
-    private val wasSignedOutInitially = !syncManager.isLoggedIn()
+    private var wasSignedOutInitially = !syncManager.isLoggedIn()
 
     val shouldPromptUpsellAfterApproval get() = wasSignedOutInitially && settings.cachedSubscription.value == null
 
@@ -54,6 +67,20 @@ class DeviceApproveViewModel @Inject constructor(
 
     fun onSetupAccountTapped() {
         eventHorizon.track(DeviceSetupAccountTappedEvent)
+    }
+
+    suspend fun switchAccount() {
+        LogBuffer.i(LogBuffer.TAG_BACKGROUND_TASKS, "User requested to switch accounts while approving a device")
+        userManager.signOutAndClearData(
+            playbackManager = playbackManager,
+            upNextQueue = upNextQueue,
+            folderManager = folderManager,
+            searchHistoryManager = searchHistoryManager,
+            episodeManager = episodeManager,
+            wasInitiatedByUser = true,
+        )?.join()
+        wasSignedOutInitially = true
+        refreshAccountState()
     }
 
     fun onDismissed() {