+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 *   New Features
     *   Log in to the Wear OS app by scanning a QR code with your phone or entering a code on the web
         ([#5964](https://github.com/Automattic/pocket-casts-android/pull/5964))
+    *   Switch accounts from the screen that appears when you pair a device by scanning its QR code
+        ([#6019](https://github.com/Automattic/pocket-casts-android/pull/6019))
 *   Bug Fixes
     *   Show every matching folder in search results instead of only the last one
         ([#5953](https://github.com/Automattic/pocket-casts-android/pull/5953))